requirements.txt
django==2.1.1
django-storages==1.7.1
google-cloud-storage==1.11.0
mysqlclient==1.3.13
honcho==1.0.1
gunicorn==19.9.0
django-htmlmin==0.10.0
django-image-cropping==1.1.0
easy_thumbnails==2.5
WhiteNoise==3.3.1
django-redis==4.9.0
django-ckeditor==5.6.1
Pillow==5.2.0